Take a viewBox object
The `viewBox` object is the container the contextmenu element is
relative, or will be positioned relative to. Rather than previous basic
assumptions of `Window`.
The viewBox size is implicitly evaluated&mdash;taken out of the end
user's scope. `viewBox` can be of type `HTMLElement` or `Window`.

This aids a better positioning system.